graylistx replaces whitelistx, better testing, also black/white agnostic
