+++ authn/webapi/interfaces/graylistx/constants/constants.go
@@ -1,0 +1,22 @@
+package constants
+
+import "time"
+
+type WhitelistxConstants struct {
+	Host        string
+	Port        int
+	Protocol    string
+	Expire      string
+	MaxIdle     int
+	IdleTimeout time.Duration
+	MaxActive   int
+}
+
+const (
+	Ok   = "OK"
+	Set  = "SET"
+	Get  = "GET"
+	Del  = "DEL"
+	Px   = "PX"
+)
+
